Rudimentary Core Shell Dual Channel
I have implemented dual channels for the CFODs, so that when they get
created they have a 'False' (down) channel. When these are calculated,
we just create some special rule depending on the type of FOD (CFOD,
BFOD, and FFOD). In it's most primitive form, this could simply repeat
the BFODs and FFODs in their place, though we could also create a very
simple system to alternate the DOWN (False) FODs to be rotate by 90
degrees in DBFOD and 30 degrees in TBFODs.